+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 # Changelog
 
 ## Pre release
+* KLS-52 - Three columns block for Article page
 
 ### Bugs fixed
 
 ### Hotfix
 
 ### Enhancements
+* KLS-54 - Add subscribe to FTA updates form
 
 ## [2.16.2](https://github.com/uktrade/great-cms/releases/tag/2.16.0)
 ### Enhancements